// app/requests/_personal/getBik/gitBikSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetBikSuccess {
  result: {
    bankData: {
      value: string;
      bic: string;
      label: string;
      disabled: boolean;
    }[];
  };
  status: "success";
}